+++ front-end/src/components/Map.js
@@ -3,6 +3,7 @@ import styled from "styled-components";
 const StyledMap = styled.div`
     width: 100%;
     height: 100vh;
+    z-index: 2;
 `;
 
 function Map() {
