+++ front-end/src/routes/Path/PathPresenter.js
@@ -4,6 +4,7 @@ import PropTypes from "prop-types";
 import styled from "styled-components";
 import Map from "../../components/Map";
 
+
 function PathPresenter(props) {
   
   return (
